moving tests to separate unit and acceptance packages
