--- src/test/java/com/spartaglobal/moviesapi/CucumberTestContextConfiguration.java
@@ -1,19 +1,0 @@
-package com.spartaglobal.moviesapi;
-
-import io.cucumber.junit.platform.engine.Cucumber;
-import io.cucumber.spring.CucumberContextConfiguration;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
-
-@CucumberContextConfiguration
-@SpringBootTest
-@AutoConfigureMockMvc
-@Sql(scripts = {"classpath:schema.sql", "classpath:data.sql"},
-    executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
-@ActiveProfiles("dev")
-public class CucumberTestContextConfiguration {
-
-}
