+++ src/test/java/com/spartaglobal/moviesapi/acceptance/RunCucumberTest.java
@@ -1,4 +1,4 @@
-package com.spartaglobal.moviesapi;
+package com.spartaglobal.moviesapi.acceptance;
 
 
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
@@ -7,11 +7,10 @@ import org.junit.platform.suite.api.ConfigurationParameter;
 import org.junit.platform.suite.api.IncludeEngines;
 import org.junit.platform.suite.api.SelectClasspathResource;
 import org.junit.platform.suite.api.Suite;
-import org.springframework.test.context.ActiveProfiles;
 
 @Suite
 @IncludeEngines("cucumber")
-@SelectClasspathResource("com/spartaglobal/moviesapi/steps")
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.spartaglobal.moviesapi.steps")
+@SelectClasspathResource("com/spartaglobal/moviesapi/acceptance/steps")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.spartaglobal.moviesapi.acceptance.steps")
 public class RunCucumberTest {
 }
